feat(地区、商铺): 完成商铺接口对接
1.完成地区树设计
2.数据表字段修改
3.商铺接口对接

+++ src/main/java/com/lsd/fun/modules/cos/controller/COSController.java
@@ -38,7 +38,7 @@ public class COSController {
 
     @ApiOperation("删除文件接口")
     @PostMapping("/delete")
-    public R deleteFile(@ApiParam(value = "需要删除的文件id") @RequestParam Long tFileId) {
+    public R deleteFile(@ApiParam(value = "需要删除的文件id") @RequestParam Integer tFileId) {
         tFileService.deleteById(tFileId);
         return R.ok();
     }
